improving the tabbed interface. see full commit for problems...
Because the Google Map is in the jQuery tabs, some UI problems may occur.
The InfoWindow stopped working, but does work for what is the default
(popups on origin and destination).

Because of limited screen space, putting the input form in a tab makes
sense. Adding functionality to the existing parts of PlaceMap shouldn't be a
problem.