updated css to make the tabs near 100% of window
the problem lies in that the container div needed to be absolute position.
it's position could be automatic but needed to be absolute. the tab content
is still set to 80% to prevent runoff and the body's overflow is set to
hidden due to the scrollbars causing runoff. there is no 'fill in the rest
of the div' in css as far as i know and the tab button heights may be hard
to get at.